adjust test case for ci

+++ test.sh
@@ -2,15 +2,17 @@
 set -e
 echo "" > coverage.txt
 
-GOMAXPROCS=1 go test -timeout 900s `go list ./... | grep -v consistence | grep -v nsqadmin`
-#GOMAXPROCS=4 go test -timeout 900s -race `go list ./... | grep -v consistence | grep -v nsqadmin`
-for d in $(go list ./... | grep -v consistence | grep -v nsqadmin); do
-    GOMAXPROCS=4 go test -timeout 900s -race -coverprofile=profile.out -covermode=atomic $d
-    if [ -f profile.out ]; then
-        cat profile.out >> coverage.txt
-        rm profile.out
-    fi
-done
+if [ "$TEST_RACE" = "false" ]; then
+    GOMAXPROCS=1 go test -timeout 900s `go list ./... | grep -v consistence | grep -v nsqadmin`
+else
+    for d in $(go list ./... | grep -v consistence | grep -v nsqadmin); do
+        GOMAXPROCS=4 go test -timeout 900s -race -coverprofile=profile.out -covermode=atomic $d
+        if [ -f profile.out ]; then
+            cat profile.out >> coverage.txt
+            rm profile.out
+        fi
+    done
+fi
 
 # no tests, but a build is something
 for dir in $(find apps bench -maxdepth 1 -type d) nsqadmin; do
